StudentMy-master/requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.3
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-npm==1.0.0
django-widget-tweaks==1.4.1
gunicorn==19.7.1
idna==2.6
Pillow==5.0.0
psycopg2==2.7.4
python-decouple==3.1
python-firebase==1.2
pytz==2018.3
raven==6.6.0
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
Unipath==1.1
urllib3==1.22
whitenoise==3.3.1
